requirements.txt
attrs==24.3.0
build==1.2.2.post1
amqp==5.3.1
certifi==2024.12.14
charset-normalizer==3.4.1
asgiref==3.8.1
cattrs==24.1.2
dill==0.3.9
async-timeout==5.0.1
idna==3.10
flatbuffers==24.3.25
numpy==1.26.4
billiard==4.2.1
exceptiongroup==1.2.2
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
openmeteo_requests==1.3.0
openmeteo_sdk==1.18.6
click-plugins==1.1.1
click-repl==0.3.0
pip-tools==7.4.1
pyproject_hooks==1.2.0
colorama==0.4.6
Django==5.0.11
platformdirs==4.3.6
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-environ==0.12.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
kombu==5.4.2
packaging==24.2
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
pandas==2.2.2
requests-cache==1.2.1
requests==2.32.3
retry-requests==2.0.0
scikit-learn==1.6.0
tomli==2.2.1
url-normalize==1.4.3
urllib3==2.3.0
reportlab==4.3.1
matplotlib==3.8.0
pypng==0.20220715.0
qrcode==7.4.2
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
phonenumbers>=8.13.50
django-axes
django-defender
django-recaptcha
django-jazzmin

